Add files via upload
This project develops a Face Mask Detection system trained on a custom dataset to accurately detect individuals wearing or not wearing masks in real-time. It employs deep learning techniques, specifically convolutional neural networks (CNNs), to enhance public health and safety measures. The system is designed to be deployed in various environments, such as security cameras and access control systems, to ensure adherence to mask-wearing protocols effectively.